refactor: improve namings, docs and structure
