+++ src/components/index.ts
@@ -11,7 +11,6 @@ export { EntityDetail } from './EntityDetail/EntityDetail';
 export { EntityTitle } from './EntityTitle/EntityTitle';
 export { HideContent } from './HideContent/HideContent';
 export { TextButton } from './TextButton/TextButton';
-export { ViewSettingsForm } from './ViewSettingsForm/ViewSettingsForm';
 export { SafeImage } from './SafeImage/SafeImage';
 export { RadioGroup } from './RadioGroup/RadioGroup';
 export { EntityTabs } from './EntityTabs/EntityTabs';
