+++ src/models/pagination.ts
@@ -4,7 +4,7 @@ export interface Pagination<T> {
   /** Items on the page. */
   readonly items: readonly T[];
 
-  /** Whether the pagination s next page or not. */
+  /** Whether the pagination has next page or not. */
   readonly hasMore: boolean;
 
   /** Max amount of items on the page. */
