+++ src/utils/debounce.ts
@@ -4,21 +4,21 @@ export namespace Debounce {
   /**
    * Decorator that applies delay to the execution of a function.
    * @param fn The source function to which delay will be applied.
-   * @param delay Delay of the execution of the function in ms.
+   * @param delayTime Delay of the execution of the function in milliseconds.
    */
   export function apply<TArgs extends unknown[]>(
     fn: (this: void, ...args: TArgs) => unknown,
-    delay = DEFAULT_DEBOUNCE_TIME_IN_MILLISECONDS,
+    delayTime = DEFAULT_DEBOUNCE_TIME_IN_MILLISECONDS,
   ): (...args: TArgs) => void {
-    let timeoutID: number | null = null;
+    let timeoutId: number | null = null;
     return (...args: TArgs) => {
-      if (timeoutID != null) {
-        window.clearTimeout(timeoutID);
+      if (timeoutId != null) {
+        window.clearTimeout(timeoutId);
       }
 
-      timeoutID = window.setTimeout(() => {
+      timeoutId = window.setTimeout(() => {
         fn(...args);
-      }, delay);
+      }, delayTime);
     };
   }
 }
